fix instant converters with time zones and formatting
