+++ src/main/java/ca/bc/gov/open/Scss/Configuration/SoapConfig.java
@@ -1,6 +1,7 @@
 package ca.bc.gov.open.Scss.Configuration;
 
 import ca.bc.gov.open.Scss.Models.Serializers.InstantDeserializer;
+import ca.bc.gov.open.Scss.Models.Serializers.InstantSerializer;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
@@ -63,6 +64,7 @@ public class SoapConfig extends WsConfigurerAdapter {
         objectMapper.disable(DeserializationFeature.ADJUST_DATES_TO_CONTEXT_TIME_ZONE);
         SimpleModule module = new SimpleModule();
         module.addDeserializer(Instant.class, new InstantDeserializer());
+        module.addSerializer(Instant.class, new InstantSerializer());
         objectMapper.registerModule(module);
         return objectMapper;
     }
