+++ src/main/java/ca/bc/gov/open/Scss/Models/Serializers/InstantSerializer.java
@@ -1,34 +1,23 @@
 package ca.bc.gov.open.Scss.Models.Serializers;
 
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.databind.JsonSerializer;
+import com.fasterxml.jackson.databind.SerializerProvider;
+import java.io.IOException;
 import java.time.Instant;
-import java.util.Date;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
 import java.util.Locale;
 
-public final class InstantSerializer {
-
-    private InstantSerializer() {}
-
-    public static String print(Instant xml) {
-        String first = xml.toString();
-        return first.substring(0, first.length() - 1);
-    }
-
-    public static Instant parse(String value) {
-        try {
-            Date d;
-            // Try to parse a datetime first then try date only if both fail return null
-            try {
-                // Date time parser
-                d = new SimpleDateFormat("dd-MMM-yy hh.mm.ss.SSSSSS a", Locale.US).parse(value);
-            } catch (ParseException ex) {
-                // Date only parser
-                d = new SimpleDateFormat("dd-MMM-yy", Locale.US).parse(value);
-            }
-            return d.toInstant();
-        } catch (Exception ex) {
-            return null;
-        }
+public class InstantSerializer extends JsonSerializer<Instant> {
+    @Override
+    public void serialize(Instant value, JsonGenerator gen, SerializerProvider serializers)
+            throws IOException {
+        String out =
+                DateTimeFormatter.ofPattern("dd-MMM-yy")
+                        .withZone(ZoneId.of("GMT-7"))
+                        .withLocale(Locale.US)
+                        .format(value);
+        gen.writeString(out);
     }
 }
