+++ src/main/resources/jaxbConfigs/jaxb-datetime-binding.xjb
@@ -2,7 +2,7 @@
         <jaxb:globalBindings>
             <jaxb:serializable uid="1" />
             <jaxb:javaType name="java.time.Instant" xmlType="xsd:dateTime"
-                   parseMethod="ca.bc.gov.open.Scss.Models.Serializers.InstantSerializer.parse"
-                   printMethod="ca.bc.gov.open.Scss.Models.Serializers.InstantSerializer.print" />
+                   parseMethod="ca.bc.gov.open.Scss.Models.Serializers.InstantSoapConverter.parse"
+                   printMethod="ca.bc.gov.open.Scss.Models.Serializers.InstantSoapConverter.print" />
         </jaxb:globalBindings>
 </jaxb:bindings>
